move tests to mod.rs
