+++ crates/bevy_ecs/src/schedule_v3/mod.rs
@@ -19,9 +19,12 @@ pub use self::state::*;
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate as bevy_ecs;
-    use crate::system::*;
-    use crate::world::World;
+    use std::sync::atomic::{AtomicU32, Ordering};
+
+    pub use crate as bevy_ecs;
+    pub use crate::schedule_v3::{IntoSystemConfig, IntoSystemSetConfig, Schedule, SystemSet};
+    pub use crate::system::{Res, ResMut};
+    pub use crate::{prelude::World, system::Resource};
 
     #[allow(dead_code)]
     #[derive(ScheduleLabel, Clone, Debug, PartialEq, Eq, Hash)]
@@ -42,185 +45,415 @@ mod tests {
         X,
     }
 
-    #[test]
-    fn correct_order() {
-        #[derive(Resource)]
-        struct X(Vec<TestSet>);
+    #[derive(Resource, Default)]
+    struct SystemOrder(Vec<u32>);
 
-        let mut world = World::new();
-        world.insert_resource(X(Vec::new()));
+    #[derive(Resource, Default)]
+    struct RunConditionBool(pub bool);
 
-        fn run(set: TestSet) -> impl FnMut(ResMut<X>) {
-            move |mut x: ResMut<X>| {
-                x.0.push(set.clone());
-            }
+    #[derive(Resource, Default)]
+    struct Counter(pub AtomicU32);
+
+    fn make_exclusive_system(tag: u32) -> impl FnMut(&mut World) {
+        move |world| world.resource_mut::<SystemOrder>().0.push(tag)
+    }
+
+    fn make_function_system(tag: u32) -> impl FnMut(ResMut<SystemOrder>) {
+        move |mut resource: ResMut<SystemOrder>| resource.0.push(tag)
+    }
+
+    fn named_system(mut resource: ResMut<SystemOrder>) {
+        resource.0.push(u32::MAX);
+    }
+
+    fn named_exclusive_system(world: &mut World) {
+        world.resource_mut::<SystemOrder>().0.push(u32::MAX);
+    }
+
+    fn counting_system(counter: Res<Counter>) {
+        counter.0.fetch_add(1, Ordering::Relaxed);
+    }
+
+    mod system_execution {
+        use std::sync::{Arc, Barrier};
+
+        use super::*;
+
+        #[test]
+        fn run_system() {
+            let mut world = World::default();
+            let mut schedule = Schedule::default();
+
+            world.init_resource::<SystemOrder>();
+
+            schedule.add_system(make_function_system(0));
+            schedule.run(&mut world);
+
+            assert_eq!(world.resource::<SystemOrder>().0, vec![0]);
         }
 
-        let mut schedule = Schedule::new();
-        schedule.add_systems(
-            (
-                run(TestSet::A),
-                run(TestSet::B),
-                run(TestSet::C),
-                run(TestSet::D),
-            )
-                .chain(),
-        );
+        #[test]
+        fn run_exclusive_system() {
+            let mut world = World::default();
+            let mut schedule = Schedule::default();
 
-        schedule.run(&mut world);
-        let X(results) = world.remove_resource::<X>().unwrap();
-        assert_eq!(
-            results,
-            vec![TestSet::A, TestSet::B, TestSet::C, TestSet::D]
-        );
+            world.init_resource::<SystemOrder>();
+
+            schedule.add_system(make_exclusive_system(0));
+            schedule.run(&mut world);
+
+            assert_eq!(world.resource::<SystemOrder>().0, vec![0]);
+        }
+
+        #[test]
+        fn parallel_execution() {
+            let mut world = World::default();
+            let mut schedule = Schedule::default();
+
+            let barrier = Arc::new(Barrier::new(3));
+
+            let barrier1 = barrier.clone();
+            schedule.add_system(move || {
+                barrier1.wait();
+            });
+            let barrier2 = barrier.clone();
+            schedule.add_system(move || {
+                barrier2.wait();
+            });
+            schedule.add_system(move || {
+                barrier.wait();
+            });
+
+            schedule.run(&mut world);
+        }
     }
 
-    #[test]
-    #[should_panic]
-    fn dependency_loop() {
-        let mut schedule = Schedule::new();
-        schedule.configure_set(TestSet::X.after(TestSet::X));
+    mod system_ordering {
+        use super::*;
+
+        #[test]
+        fn order_systems() {
+            let mut world = World::default();
+            let mut schedule = Schedule::default();
+
+            world.init_resource::<SystemOrder>();
+
+            schedule.add_system(named_system);
+            schedule.add_system(make_function_system(1).before(named_system));
+            schedule.add_system(make_function_system(0).after(named_system).in_set(TestSet::A));
+            schedule.run(&mut world);
+
+            assert_eq!(world.resource::<SystemOrder>().0, vec![1, u32::MAX, 0]);
+
+            world.insert_resource(SystemOrder::default());
+
+            assert_eq!(world.resource::<SystemOrder>().0, vec![]);
+
+            // modify the schedule after it's been initialized and test ordering with sets
+            schedule.configure_set(TestSet::A.after(named_system));
+            schedule.add_system(make_function_system(3).before(TestSet::A).after(named_system));
+            schedule.add_system(make_function_system(4).after(TestSet::A));
+            schedule.run(&mut world);
+
+            assert_eq!(
+                world.resource::<SystemOrder>().0,
+                vec![1, u32::MAX, 3, 0, 4]
+            );
+        }
+
+        #[test]
+        fn order_exclusive_systems() {
+            let mut world = World::default();
+            let mut schedule = Schedule::default();
+
+            world.init_resource::<SystemOrder>();
+
+            schedule.add_system(named_exclusive_system);
+            schedule.add_system(make_exclusive_system(1).before(named_exclusive_system));
+            schedule.add_system(make_exclusive_system(0).after(named_exclusive_system));
+            schedule.run(&mut world);
+
+            assert_eq!(world.resource::<SystemOrder>().0, vec![1, u32::MAX, 0]);
+        }
+
+        #[test]
+        fn add_systems_correct_order() {
+            #[derive(Resource)]
+            struct X(Vec<TestSet>);
+    
+            let mut world = World::new();
+            world.init_resource::<SystemOrder>();
+    
+            let mut schedule = Schedule::new();
+            schedule.add_systems(
+                (
+                    make_function_system(0),
+                    make_function_system(1),
+                    make_exclusive_system(2),
+                    make_function_system(3),
+                )
+                    .chain(),
+            );
+    
+            schedule.run(&mut world);
+            assert_eq!(world.resource::<SystemOrder>().0, vec![0, 1, 2, 3]);
+        }
     }
 
-    #[test]
-    fn dependency_cycle() {
-        let mut world = World::new();
-        let mut schedule = Schedule::new();
+    mod run_conditions {
+        use super::*;
 
-        schedule.configure_set(TestSet::A.after(TestSet::B));
-        schedule.configure_set(TestSet::B.after(TestSet::A));
+        #[test]
+        fn system_with_run_condition() {
+            let mut world = World::default();
+            let mut schedule = Schedule::default();
 
-        let result = schedule.initialize(&mut world);
-        assert!(matches!(result, Err(BuildError::DependencyCycle)));
+            world.init_resource::<RunConditionBool>();
+            world.init_resource::<SystemOrder>();
 
-        fn foo() {}
-        fn bar() {}
+            schedule.add_system(
+                make_function_system(0).run_if(|condition: Res<RunConditionBool>| condition.0),
+            );
 
-        let mut world = World::new();
-        let mut schedule = Schedule::new();
+            schedule.run(&mut world);
+            assert_eq!(world.resource::<SystemOrder>().0, vec![]);
 
-        schedule.add_systems((foo.after(bar), bar.after(foo)));
-        let result = schedule.initialize(&mut world);
-        assert!(matches!(result, Err(BuildError::DependencyCycle)));
+            world.resource_mut::<RunConditionBool>().0 = true;
+            schedule.run(&mut world);
+            assert_eq!(world.resource::<SystemOrder>().0, vec![0]);
+        }
+
+        #[test]
+        fn run_exclusive_system_with_run_condition() {
+            let mut world = World::default();
+            let mut schedule = Schedule::default();
+
+            world.init_resource::<RunConditionBool>();
+            world.init_resource::<SystemOrder>();
+
+            schedule.add_system(
+                make_exclusive_system(0).run_if(|condition: Res<RunConditionBool>| condition.0),
+            );
+
+            schedule.run(&mut world);
+            assert_eq!(world.resource::<SystemOrder>().0, vec![]);
+
+            world.resource_mut::<RunConditionBool>().0 = true;
+            schedule.run(&mut world);
+            assert_eq!(world.resource::<SystemOrder>().0, vec![0]);
+        }
+
+        #[test]
+        fn multiple_run_criteria_on_system() {
+            let mut world = World::default();
+            let mut schedule = Schedule::default();
+
+            world.init_resource::<Counter>();
+
+            schedule.add_system(counting_system.run_if(|| false).run_if(|| false));
+            schedule.add_system(counting_system.run_if(|| true).run_if(|| false));
+            schedule.add_system(counting_system.run_if(|| false).run_if(|| true));
+            schedule.add_system(counting_system.run_if(|| true).run_if(|| true));
+
+            schedule.run(&mut world);
+            assert_eq!(world.resource::<Counter>().0.load(Ordering::Relaxed), 1);
+        }
+
+        #[test]
+        fn multiple_run_criteria_on_system_sets() {
+            let mut world = World::default();
+            let mut schedule = Schedule::default();
+
+            world.init_resource::<Counter>();
+
+            schedule.configure_set(TestSet::A.run_if(|| false).run_if(|| false));
+            schedule.add_system(counting_system.in_set(TestSet::A));
+            schedule.configure_set(TestSet::B.run_if(|| true).run_if(|| false));
+            schedule.add_system(counting_system.in_set(TestSet::B));
+            schedule.configure_set(TestSet::C.run_if(|| false).run_if(|| true));
+            schedule.add_system(counting_system.in_set(TestSet::C));
+            schedule.configure_set(TestSet::D.run_if(|| true).run_if(|| true));
+            schedule.add_system(counting_system.in_set(TestSet::D));
+
+            schedule.run(&mut world);
+            assert_eq!(world.resource::<Counter>().0.load(Ordering::Relaxed), 1);
+        }
+
+        #[test]
+        fn systems_nested_in_system_sets() {
+            let mut world = World::default();
+            let mut schedule = Schedule::default();
+
+            world.init_resource::<Counter>();
+
+            schedule.configure_set(TestSet::A.run_if(|| false));
+            schedule.add_system(counting_system.in_set(TestSet::A).run_if(|| false));
+            schedule.configure_set(TestSet::B.run_if(|| true));
+            schedule.add_system(counting_system.in_set(TestSet::B).run_if(|| false));
+            schedule.configure_set(TestSet::C.run_if(|| false));
+            schedule.add_system(counting_system.in_set(TestSet::C).run_if(|| true));
+            schedule.configure_set(TestSet::D.run_if(|| true));
+            schedule.add_system(counting_system.in_set(TestSet::D).run_if(|| true));
+
+            schedule.run(&mut world);
+            assert_eq!(world.resource::<Counter>().0.load(Ordering::Relaxed), 1);
+        }
     }
 
-    #[test]
-    #[should_panic]
-    fn hierarchy_loop() {
-        let mut schedule = Schedule::new();
-        schedule.configure_set(TestSet::X.in_set(TestSet::X));
+    mod schedule_build_errors {
+        use super::*;
+    
+        #[test]
+        #[should_panic]
+        fn dependency_loop() {
+            let mut schedule = Schedule::new();
+            schedule.configure_set(TestSet::X.after(TestSet::X));
+        }
+    
+        #[test]
+        fn dependency_cycle() {
+            let mut world = World::new();
+            let mut schedule = Schedule::new();
+    
+            schedule.configure_set(TestSet::A.after(TestSet::B));
+            schedule.configure_set(TestSet::B.after(TestSet::A));
+    
+            let result = schedule.initialize(&mut world);
+            assert!(matches!(result, Err(BuildError::DependencyCycle)));
+    
+            fn foo() {}
+            fn bar() {}
+    
+            let mut world = World::new();
+            let mut schedule = Schedule::new();
+    
+            schedule.add_systems((foo.after(bar), bar.after(foo)));
+            let result = schedule.initialize(&mut world);
+            assert!(matches!(result, Err(BuildError::DependencyCycle)));
+        }
+    
+        #[test]
+        #[should_panic]
+        fn hierarchy_loop() {
+            let mut schedule = Schedule::new();
+            schedule.configure_set(TestSet::X.in_set(TestSet::X));
+        }
+    
+        #[test]
+        fn hierarchy_cycle() {
+            let mut world = World::new();
+            let mut schedule = Schedule::new();
+    
+            schedule.configure_set(TestSet::A.in_set(TestSet::B));
+            schedule.configure_set(TestSet::B.in_set(TestSet::A));
+    
+            let result = schedule.initialize(&mut world);
+            assert!(matches!(result, Err(BuildError::HierarchyCycle)));
+        }
+    
+        #[test]
+        fn system_type_set_ambiguity() {
+            // Define some systems.
+            fn foo() {}
+            fn bar() {}
+    
+            let mut world = World::new();
+            let mut schedule = Schedule::new();
+    
+            // Schedule `bar` to run after `foo`.
+            schedule.add_system(foo);
+            schedule.add_system(bar.after(foo));
+    
+            // There's only one `foo`, so it's fine.
+            let result = schedule.initialize(&mut world);
+            assert!(result.is_ok());
+    
+            // Schedule another `foo`.
+            schedule.add_system(foo);
+    
+            // When there are multiple instances of `foo`, dependencies on
+            // `foo` are no longer allowed. Too much ambiguity.
+            let result = schedule.initialize(&mut world);
+            assert!(matches!(result, Err(BuildError::SystemTypeSetAmbiguity(_))));
+    
+            // same goes for `ambiguous_with`
+            let mut schedule = Schedule::new();
+            schedule.add_system(foo);
+            schedule.add_system(bar.ambiguous_with(foo));
+            let result = schedule.initialize(&mut world);
+            assert!(result.is_ok());
+            schedule.add_system(foo);
+            let result = schedule.initialize(&mut world);
+            assert!(matches!(result, Err(BuildError::SystemTypeSetAmbiguity(_))));
+        }
+    
+        #[test]
+        #[should_panic]
+        fn in_system_type_set() {
+            fn foo() {}
+            fn bar() {}
+    
+            let mut schedule = Schedule::new();
+            schedule.add_system(foo.in_set(bar.into_system_set()));
+        }
+    
+        #[test]
+        #[should_panic]
+        fn configure_system_type_set() {
+            fn foo() {}
+            let mut schedule = Schedule::new();
+            schedule.configure_set(foo.into_system_set());
+        }
+    
+        #[test]
+        fn hierarchy_conflict() {
+            let mut world = World::new();
+            let mut schedule = Schedule::new();
+    
+            // Add `A`.
+            schedule.configure_set(TestSet::A);
+    
+            // Add `B` as child of `A`.
+            schedule.configure_set(TestSet::B.in_set(TestSet::A));
+    
+            // Add `X` as child of both `A` and `B`.
+            schedule.configure_set(TestSet::X.in_set(TestSet::A).in_set(TestSet::B));
+    
+            // `X` cannot be the `A`'s child and grandchild at the same time.
+            let result = schedule.initialize(&mut world);
+            assert!(matches!(result, Err(BuildError::HierarchyConflict)));
+        }
+    
+        #[test]
+        fn cross_dependency() {
+            let mut world = World::new();
+            let mut schedule = Schedule::new();
+    
+            // Add `B` and give it both kinds of relationships with `A`.
+            schedule.configure_set(TestSet::B.in_set(TestSet::A));
+            schedule.configure_set(TestSet::B.after(TestSet::A));
+            let result = schedule.initialize(&mut world);
+            assert!(matches!(result, Err(BuildError::CrossDependency(_, _))));
+        }
+    
+        #[test]
+        fn ambiguity() {
+            #[derive(Resource)]
+            struct X;
+    
+            fn res_ref(_x: Res<X>) {}
+            fn res_mut(_x: ResMut<X>) {}
+    
+            let mut world = World::new();
+            let mut schedule = Schedule::new();
+    
+            schedule.add_systems((res_ref, res_mut));
+            let result = schedule.initialize(&mut world);
+            assert!(matches!(result, Err(BuildError::Ambiguity)));
+        }
     }
 
-    #[test]
-    fn hierarchy_cycle() {
-        let mut world = World::new();
-        let mut schedule = Schedule::new();
-
-        schedule.configure_set(TestSet::A.in_set(TestSet::B));
-        schedule.configure_set(TestSet::B.in_set(TestSet::A));
-
-        let result = schedule.initialize(&mut world);
-        assert!(matches!(result, Err(BuildError::HierarchyCycle)));
-    }
-
-    #[test]
-    fn system_type_set_ambiguity() {
-        // Define some systems.
-        fn foo() {}
-        fn bar() {}
-
-        let mut world = World::new();
-        let mut schedule = Schedule::new();
-
-        // Schedule `bar` to run after `foo`.
-        schedule.add_system(foo);
-        schedule.add_system(bar.after(foo));
-
-        // There's only one `foo`, so it's fine.
-        let result = schedule.initialize(&mut world);
-        assert!(result.is_ok());
-
-        // Schedule another `foo`.
-        schedule.add_system(foo);
-
-        // When there are multiple instances of `foo`, dependencies on
-        // `foo` are no longer allowed. Too much ambiguity.
-        let result = schedule.initialize(&mut world);
-        assert!(matches!(result, Err(BuildError::SystemTypeSetAmbiguity(_))));
-
-        // same goes for `ambiguous_with`
-        let mut schedule = Schedule::new();
-        schedule.add_system(foo);
-        schedule.add_system(bar.ambiguous_with(foo));
-        let result = schedule.initialize(&mut world);
-        assert!(result.is_ok());
-        schedule.add_system(foo);
-        let result = schedule.initialize(&mut world);
-        assert!(matches!(result, Err(BuildError::SystemTypeSetAmbiguity(_))));
-    }
-
-    #[test]
-    #[should_panic]
-    fn in_system_type_set() {
-        fn foo() {}
-        fn bar() {}
-
-        let mut schedule = Schedule::new();
-        schedule.add_system(foo.in_set(bar.into_system_set()));
-    }
-
-    #[test]
-    #[should_panic]
-    fn configure_system_type_set() {
-        fn foo() {}
-        let mut schedule = Schedule::new();
-        schedule.configure_set(foo.into_system_set());
-    }
-
-    #[test]
-    fn hierarchy_conflict() {
-        let mut world = World::new();
-        let mut schedule = Schedule::new();
-
-        // Add `A`.
-        schedule.configure_set(TestSet::A);
-
-        // Add `B` as child of `A`.
-        schedule.configure_set(TestSet::B.in_set(TestSet::A));
-
-        // Add `X` as child of both `A` and `B`.
-        schedule.configure_set(TestSet::X.in_set(TestSet::A).in_set(TestSet::B));
-
-        // `X` cannot be the `A`'s child and grandchild at the same time.
-        let result = schedule.initialize(&mut world);
-        assert!(matches!(result, Err(BuildError::HierarchyConflict)));
-    }
-
-    #[test]
-    fn cross_dependency() {
-        let mut world = World::new();
-        let mut schedule = Schedule::new();
-
-        // Add `B` and give it both kinds of relationships with `A`.
-        schedule.configure_set(TestSet::B.in_set(TestSet::A));
-        schedule.configure_set(TestSet::B.after(TestSet::A));
-        let result = schedule.initialize(&mut world);
-        assert!(matches!(result, Err(BuildError::CrossDependency(_, _))));
-    }
-
-    #[test]
-    fn ambiguity() {
-        #[derive(Resource)]
-        struct X;
-
-        fn res_ref(_x: Res<X>) {}
-        fn res_mut(_x: ResMut<X>) {}
-
-        let mut world = World::new();
-        let mut schedule = Schedule::new();
-
-        schedule.add_systems((res_ref, res_mut));
-        let result = schedule.initialize(&mut world);
-        assert!(matches!(result, Err(BuildError::Ambiguity)));
+    mod system_ambiguity_errors {
+        // TODO
     }
 }
