+++ crates/bevy_ecs/src/schedule_v3/schedule.rs
@@ -1013,261 +1013,3 @@ pub enum BuildError {
     #[error("schedule not initialized")]
     Uninitialized,
 }
-
-#[cfg(test)]
-mod tests {
-    use std::sync::atomic::{AtomicU32, Ordering};
-
-    pub use crate as bevy_ecs;
-    pub use crate::schedule_v3::{IntoSystemConfig, IntoSystemSetConfig, Schedule, SystemSet};
-    pub use crate::system::{Res, ResMut};
-    pub use crate::{prelude::World, system::Resource};
-
-    #[derive(SystemSet, Clone, Debug, PartialEq, Eq, Hash)]
-    enum Set {
-        A,
-        B,
-        C,
-        D,
-    }
-
-    #[derive(Resource, Default)]
-    struct SystemOrder(Vec<u32>);
-
-    #[derive(Resource, Default)]
-    struct RunConditionBool(pub bool);
-
-    #[derive(Resource, Default)]
-    struct Counter(pub AtomicU32);
-
-    fn make_exclusive_system(tag: u32) -> impl FnMut(&mut World) {
-        move |world| world.resource_mut::<SystemOrder>().0.push(tag)
-    }
-
-    fn make_function_system(tag: u32) -> impl FnMut(ResMut<SystemOrder>) {
-        move |mut resource: ResMut<SystemOrder>| resource.0.push(tag)
-    }
-
-    fn named_system(mut resource: ResMut<SystemOrder>) {
-        resource.0.push(u32::MAX);
-    }
-
-    fn named_exclusive_system(world: &mut World) {
-        world.resource_mut::<SystemOrder>().0.push(u32::MAX);
-    }
-
-    fn counting_system(counter: Res<Counter>) {
-        counter.0.fetch_add(1, Ordering::Relaxed);
-    }
-
-    mod system_execution {
-        use std::sync::{Arc, Barrier};
-
-        use super::*;
-
-        #[test]
-        fn run_system() {
-            let mut world = World::default();
-            let mut schedule = Schedule::default();
-
-            world.init_resource::<SystemOrder>();
-
-            schedule.add_system(make_function_system(0));
-            schedule.run(&mut world);
-
-            assert_eq!(world.resource::<SystemOrder>().0, vec![0]);
-        }
-
-        #[test]
-        fn run_exclusive_system() {
-            let mut world = World::default();
-            let mut schedule = Schedule::default();
-
-            world.init_resource::<SystemOrder>();
-
-            schedule.add_system(make_exclusive_system(0));
-            schedule.run(&mut world);
-
-            assert_eq!(world.resource::<SystemOrder>().0, vec![0]);
-        }
-
-        #[test]
-        fn parallel_execution() {
-            let mut world = World::default();
-            let mut schedule = Schedule::default();
-
-            let barrier = Arc::new(Barrier::new(3));
-
-            let barrier1 = barrier.clone();
-            schedule.add_system(move || {
-                barrier1.wait();
-            });
-            let barrier2 = barrier.clone();
-            schedule.add_system(move || {
-                barrier2.wait();
-            });
-            schedule.add_system(move || {
-                barrier.wait();
-            });
-
-            schedule.run(&mut world);
-        }
-    }
-
-    mod system_ordering {
-        use super::*;
-
-        #[test]
-        fn order_systems() {
-            let mut world = World::default();
-            let mut schedule = Schedule::default();
-
-            world.init_resource::<SystemOrder>();
-
-            schedule.add_system(named_system);
-            schedule.add_system(make_function_system(1).before(named_system));
-            schedule.add_system(make_function_system(0).after(named_system).in_set(Set::A));
-            schedule.run(&mut world);
-
-            assert_eq!(world.resource::<SystemOrder>().0, vec![1, u32::MAX, 0]);
-
-            world.insert_resource(SystemOrder::default());
-
-            assert_eq!(world.resource::<SystemOrder>().0, vec![]);
-
-            // modify the schedule after it's been initialized and test ordering with sets
-            schedule.configure_set(Set::A.after(named_system));
-            schedule.add_system(make_function_system(3).before(Set::A).after(named_system));
-            schedule.add_system(make_function_system(4).after(Set::A));
-            schedule.run(&mut world);
-
-            assert_eq!(
-                world.resource::<SystemOrder>().0,
-                vec![1, u32::MAX, 3, 0, 4]
-            );
-        }
-
-        #[test]
-        fn order_exclusive_systems() {
-            let mut world = World::default();
-            let mut schedule = Schedule::default();
-
-            world.init_resource::<SystemOrder>();
-
-            schedule.add_system(named_exclusive_system);
-            schedule.add_system(make_exclusive_system(1).before(named_exclusive_system));
-            schedule.add_system(make_exclusive_system(0).after(named_exclusive_system));
-            schedule.run(&mut world);
-
-            assert_eq!(world.resource::<SystemOrder>().0, vec![1, u32::MAX, 0]);
-        }
-    }
-
-    mod run_conditions {
-        use super::*;
-
-        #[test]
-        fn system_with_run_condition() {
-            let mut world = World::default();
-            let mut schedule = Schedule::default();
-
-            world.init_resource::<RunConditionBool>();
-            world.init_resource::<SystemOrder>();
-
-            schedule.add_system(
-                make_function_system(0).run_if(|condition: Res<RunConditionBool>| condition.0),
-            );
-
-            schedule.run(&mut world);
-            assert_eq!(world.resource::<SystemOrder>().0, vec![]);
-
-            world.resource_mut::<RunConditionBool>().0 = true;
-            schedule.run(&mut world);
-            assert_eq!(world.resource::<SystemOrder>().0, vec![0]);
-        }
-
-        #[test]
-        fn run_exclusive_system_with_run_condition() {
-            let mut world = World::default();
-            let mut schedule = Schedule::default();
-
-            world.init_resource::<RunConditionBool>();
-            world.init_resource::<SystemOrder>();
-
-            schedule.add_system(
-                make_exclusive_system(0).run_if(|condition: Res<RunConditionBool>| condition.0),
-            );
-
-            schedule.run(&mut world);
-            assert_eq!(world.resource::<SystemOrder>().0, vec![]);
-
-            world.resource_mut::<RunConditionBool>().0 = true;
-            schedule.run(&mut world);
-            assert_eq!(world.resource::<SystemOrder>().0, vec![0]);
-        }
-
-        #[test]
-        fn multiple_run_criteria_on_system() {
-            let mut world = World::default();
-            let mut schedule = Schedule::default();
-
-            world.init_resource::<Counter>();
-
-            schedule.add_system(counting_system.run_if(|| false).run_if(|| false));
-            schedule.add_system(counting_system.run_if(|| true).run_if(|| false));
-            schedule.add_system(counting_system.run_if(|| false).run_if(|| true));
-            schedule.add_system(counting_system.run_if(|| true).run_if(|| true));
-
-            schedule.run(&mut world);
-            assert_eq!(world.resource::<Counter>().0.load(Ordering::Relaxed), 1);
-        }
-
-        #[test]
-        fn multiple_run_criteria_on_system_sets() {
-            let mut world = World::default();
-            let mut schedule = Schedule::default();
-
-            world.init_resource::<Counter>();
-
-            schedule.configure_set(Set::A.run_if(|| false).run_if(|| false));
-            schedule.add_system(counting_system.in_set(Set::A));
-            schedule.configure_set(Set::B.run_if(|| true).run_if(|| false));
-            schedule.add_system(counting_system.in_set(Set::B));
-            schedule.configure_set(Set::C.run_if(|| false).run_if(|| true));
-            schedule.add_system(counting_system.in_set(Set::C));
-            schedule.configure_set(Set::D.run_if(|| true).run_if(|| true));
-            schedule.add_system(counting_system.in_set(Set::D));
-
-            schedule.run(&mut world);
-            assert_eq!(world.resource::<Counter>().0.load(Ordering::Relaxed), 1);
-        }
-
-        #[test]
-        fn systems_nested_in_system_sets() {
-            let mut world = World::default();
-            let mut schedule = Schedule::default();
-
-            world.init_resource::<Counter>();
-
-            schedule.configure_set(Set::A.run_if(|| false));
-            schedule.add_system(counting_system.in_set(Set::A).run_if(|| false));
-            schedule.configure_set(Set::B.run_if(|| true));
-            schedule.add_system(counting_system.in_set(Set::B).run_if(|| false));
-            schedule.configure_set(Set::C.run_if(|| false));
-            schedule.add_system(counting_system.in_set(Set::C).run_if(|| true));
-            schedule.configure_set(Set::D.run_if(|| true));
-            schedule.add_system(counting_system.in_set(Set::D).run_if(|| true));
-
-            schedule.run(&mut world);
-            assert_eq!(world.resource::<Counter>().0.load(Ordering::Relaxed), 1);
-        }
-    }
-
-    mod system_ambiguity_errors {
-        // TODO
-    }
-
-    mod schedule_build_errors {
-        // TODO errors other than system ambiguity errors
-    }
-}
